requirements.txt
discord.py==1.2.3
aiofiles==0.4.0
Pillow==6.1.0
python-dotenv==0.10.3
asgiref==3.2.1
aiosqlite==0.10.0
numpy==1.17.2
scikit-image==0.15.0